change the generated parameter name to something that is very unlikely to be used as a real symbol
Inspired by protobuf codegen. Prevents cases such as an enum mapping ISO-3166-2 where one of the symbols would be `in` and prior to this PR, the symbol name of the mapping function would also be `in` resulting in `case string(in)` simply comparing `in == in` instead of the actual symbol value.